% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSsemEngine.R
\name{export_lavaan_syntax}
\alias{export_lavaan_syntax}
\title{Export to Lavaan Syntax}
\usage{
export_lavaan_syntax(measurement_model, structural_model = NULL)
}
\arguments{
\item{measurement_model}{The measurement model list.}

\item{structural_model}{The structural model list.}
}
\description{
Translates the PLSsemEngine model into lavaan-compatible syntax
for CFA or CB-SEM validation.
}
