% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSsemEngine.R
\name{export_scores}
\alias{export_scores}
\title{Export Latent Variable Scores}
\usage{
export_scores(model, file = "latent_scores.csv")
}
\arguments{
\item{model}{An object of class 'pls_model' generated by \code{pls_sem}.}

\item{file}{Character. The file path and name for the exported CSV. Default is "latent_scores.csv".}
}
\value{
Invisibly returns the data frame of scores while writing the CSV file to disk.
}
\description{
Extracts and exports the estimated latent variable scores to a 
CSV file. A sequential 'Case' identifier is appended to improve data 
traceability and facilitate subsequent external analyses.
}
